fix: Nav placement was wrong
There was an inconcistency between the about page and al the other pages for the nav. this is because the `main` was shrunk to 70vw meaning that within the 70vw the nav was smaller.

this has been fixed by removing the nav from the main and placing it directly inside the body