feat: completed style for info card detailed mode
I have completed (for the most part) the styling for the info card. this includes what happens when you click it and that content inside. it doesn't cover any modes like font, or black and white
and does not scale with browser size yet. its a rough version so i can get in-person feedback on my work in class