feat: Created account management
I have created a system that allows only administrator accounts to update the information or delete an account entirely. These admins have the right to create other admin accounts and delete other admin accounts. Ontop of this they are able to Change the password of a account (not view it though)